Title: 사칙연산 함수, Time: 119ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 240527/사칙연산 함수/quadratic-operations-function.py
@@ -17,7 +17,7 @@ if '*' in s:
     n,m = map(int,s.split(" * "))
     print(f"{s} = {mul(n,m)}")
 elif '/' in s:
-    n,m = map(int,s.split(" / "))
+    n,m = map(float,s.split(" / "))
     print("{} = {:.0f}".format(s,div(n,m)))
 elif '+' in s:
     n,m = map(int,s.split(" + "))
